refactor(AppService): issue warnings from separate modules where possible
